+++ src/officialeye/_internal/diffobject/__init__.py
@@ -1,4 +1,4 @@
 """
-A difference object is either a static dictionary, or a dictionary-like
-This module imlements the concept of a difference object.
+A difference object is either a dictionary conforming to some schema, or a dictionary-like representation of how to change another dictionary.
+This module contains everything related to difference objects.
 """