+++ src/officialeye/_internal/error/errors/io.py
@@ -1,4 +1,5 @@
-from officialeye._internal.error.codes import ERR_IO_INVALID_PATH, ERR_IO_INVALID_SUPERVISION_ENGINE, ERR_IO_OPERATION_NOT_SUPPORTED_BY_DRIVER
+from officialeye._internal.error.codes import ERR_IO_INVALID_PATH, ERR_IO_INVALID_SUPERVISION_ENGINE, ERR_IO_OPERATION_NOT_SUPPORTED_BY_DRIVER, \
+    ERR_IO_INVALID_IMAGE
 from officialeye._internal.error.error import OEError
 from officialeye._internal.error.modules import ERR_MODULE_IO
 
@@ -25,3 +26,9 @@ class ErrIOInvalidPath(ErrIO):
     def __init__(self, while_text: str, problem_text: str, /, **kwargs):
         super().__init__(
             ERR_IO_INVALID_PATH[0], ERR_IO_INVALID_PATH[1], while_text, problem_text, **kwargs)
+
+
+class ErrIOInvalidImage(ErrIO):
+    def __init__(self, while_text: str, problem_text: str, /, **kwargs):
+        super().__init__(
+            ERR_IO_INVALID_IMAGE[0], ERR_IO_INVALID_IMAGE[1], while_text, problem_text, **kwargs)